// src/Models/content-types/fact_about_us.ts
import { IContentItem, Elements } from '@kontent-ai/delivery-sdk';
import { Sitemap } from '../taxonomies/sitemap_538125f';

/**
 * Generated by '@kontent-ai/model-generator@5.2.0'
 *
 * Fact about us
 * Id: b99ec220-0f2b-4658-a080-ff0afe92f6d1
 * Codename: fact_about_us
 */
export type FactAboutUs = IContentItem<{
  /**
   * Title (text)
   * Required: false
   * Id: 2375ca8a-43d2-7282-162c-922a6fdeba3e
   * Codename: title
   *
   * Provide a title that fits within 60 characters.
   */
  title: Elements.TextElement;

  /**
   * Description (rich_text)
   * Required: false
   * Id: 1b658e81-88c9-73d1-374d-7a60e3756ef7
   * Codename: description
   *
   * Include a main goal of our business. The limit is 80 characters.
   */
  description: Elements.RichTextElement;

  /**
   * Image (asset)
   * Required: false
   * Id: 68d194b6-efad-6b25-89a4-2dfc75fed5a5
   * Codename: image
   *
   * Attach a teaser image; max. dimensions are 1280 × 600 px; allowed formats are *.jpg, *.png, *.gif.
   */
  image: Elements.AssetsElement;

  /**
   * Sitemap (taxonomy)
   * Required: false
   * Id: 23a715f9-755c-4b9d-b6b3-e02fc1b09b83
   * Codename: sitemap
   */
  sitemap: Elements.TaxonomyElement<Sitemap>;
}>;